pyproject.toml
[project]
name = "pyplate"
version = "1.0.0"
description = "Python project template"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic==2.12.4",
    "pydantic-settings==2.11.0",
]

[dependency-groups]
dev = [
    "bandit[toml]==1.9.2",
    "prek>=0.2.30",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "ruff==0.14.14",
    "ty>=0.0.14",
    "python-dotenv==1.2.1",
]


[tool.mypy]

ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
show_error_codes = true
disable_error_code = ["prop-decorator"]

exclude = ["tests", ".venv", "venv", "data", "private", "dump"]


[tool.ruff]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black
line-length = 120
indent-width = 4

target-version = "py314"


[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
extend-select = ["I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"



[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
# minversion = "6.0"
# addopts = "-ra -q"
testpaths = [
    "tests",
]
norecursedirs = [
    "data",
]


[tool.coverage.run]
omit =  [
    "*__init__*",
    "*test*",
    "*/usr/*",
    "main.py",
    "pic_back/scripts/*"
]

[tool.coverage.report]
precision=2
fail_under=90.00

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv", ".mypy_cache", ".pytest_cache", ".ruff_cache", "dump"]
